Update class attributes, add single quotes to typing annotations in system/executor/docker/*
Implementation and name attributes of Executor class (docker/executor_docker.py)
have been changed to static method and property.

async / await keywords temporarily removed

Typing annotatios of objects that are not imported for the runtime but
serve only for static type checking must be surrounded in singlequotes
if used in method signatures.