Update utils module
Surround typing annotations with single quotes in method signatures.
Typing annotatios of objects that are not imported for the runtime but
serve only for static type checking must be surrounded in singlequotes
if used in method signatures.

Replace "'%s' % var" kinds of string formattin for f-strings

Add callable check in get_subclass_with_prop_value() method due to
necessary change in executor classes.