Add new version of command base and generic command options
Things like hooks and callbacks should be part of the command instead
of an execution. Among other reasons for this is that it should be known
before the command gets executed what should happend. Neither command itself
or the execution should be responsible for the code that should run.
That is executor's responsibility. New command base is updated based on
that believe.

Added more flexible classes related to command options. This class is
more generic and will allow for dynamic creation of new classes via
principle of decorators, but with implementation similar to Python's
Enum classes. Though, this feature haven't been added yet.